backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
python-multipart
pandas
numpy
pyarrow
langgraph
google-generativeai
docker
orjson
pytest
websockets
